+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to **CreatureChat™** are documented in this file. The form
     - Ignores chest-based mobs (horse with chest, donkey with chest, llama with chest)
     - Added Fabric DataGen for building compatible loot-tables (1.20 to 1.21.7 support)
     - Randomly assign inventory loot items into random slots
+    - Main/Off hand support for max friendship (uses bottom left 2 slots for main/off hand)
 - Validation of mixin targets (on build script)
 - Solutions to common errors are now displayed on screen (i.e. more helpful)
 - New keep alive and accept HTTP headers